+++ cli/components/accounts.js
@@ -15,7 +15,6 @@ const accounts = async (backFunction) => {
         choices: [
             `Show Accounts(${accountsList.length})`,
             'Add Account',
-            'Add Bulk Accounts',
             'Remove Account',
 
             new inquirer.Separator(), // Add a separator line
